LPS-146375 search-experiences-web: Organize Preview sidebar components together
