+++ modules/dxp/apps/search-experiences/search-experiences-web/src/main/resources/META-INF/resources/sxp_blueprint_admin/js/edit_sxp_blueprint/preview_sidebar/ResultListItem.js
@@ -17,8 +17,8 @@ import getCN from 'classnames';
 import moment from 'moment';
 import React, {useContext, useState} from 'react';
 
-import {PreviewModalWithCopyDownload} from '../shared/PreviewModal';
-import ThemeContext from '../shared/ThemeContext';
+import {PreviewModalWithCopyDownload} from '../../shared/PreviewModal';
+import ThemeContext from '../../shared/ThemeContext';
 
 const getResultDefaultKeys = (locale) => [
 	'entryClassName',
